minor code cleaning, rename one_hot_encode_data, move is_categorical file
