+++ pandas_code/code_templates/predict_decision_tree.py
@@ -2,4 +2,3 @@
 def get_code(args):
    test_preds = args[0].predict(args[1])
    return test_preds
-
