+++ pandas_code/code_templates/read_csv.py
@@ -1,4 +1,3 @@
-
 #  file into a dataframe for later processing
 import pandas as pd
 def get_code(args):
