+++ pandas_code/is_categorical.py
