+++ pandas_code/parse_template.py
@@ -15,10 +15,6 @@ def parse_template(template_name, args):
             continue
          if trimmed_line.strip().startswith("return"):
             continue
-            # this is no longer being skipped
-            # needed?
-         if trimmed_line.strip().find("is_categorical") != -1:
-            continue
          if trimmed_line.startswith("#"):
             generated_comments.append(trimmed_line)
          else:
